# Commit 1 Abstract database functions so the application doesn't need to know the specifics of the underlying database technology.
## Changes
- index.html
  - Added reference to indexeddb-handler.js - Implementation of `indexedDbHandler`.
### assets
- indexeddb-handler.js
  - Renamed db-handler.js (*Old*) to indexeddb-handler.js and changed `dbHandler` references to `indexedDbHandler`.
  - Added new function `getIndexSorted` to standardize the sorted-query contract.
- db-handler.js (*New*) -
  - Created a new db-handler.js file to serve as an abstraction of database calls from the specific database technology.
### activities
- Updated the following files to use the abstracted functions against the IndexedDB.
  - diary.js
  - foodlist.js
  - foods-meals-recipes.js
  - meal-editor.js
  - meals.js
  - recipe-editor.js
  - settings.js
  - statistics.js

Progress on #728
