+++ www/activities/statistics/js/statistics.js
@@ -597,47 +597,39 @@ app.Stats = {
     return [fromDate, toDate]
   },
 
-  getDataFromDb: function(from, to) {
-    return new Promise(async function(resolve, reject) {
-      let result = {
-        "timestamps": [],
-        "items": [],
-        "stats": []
-      };
+  getDataFromDb: async function(from, to) {
+    let result = {
+      "timestamps": [],
+      "items": [],
+      "stats": []
+    };
 
+    if (!to) {
+      to = new Date();
+    }
 
-      if (!to) {
-        to = new Date();
+    // Make date range inclusive of the whole days at either end
+    let toDate = new Date(Date.UTC(to.getFullYear(), to.getMonth(), to.getDate()));
+    toDate.setHours(0, 0, 0, 0);
+    let fromDate = new Date(Date.UTC(from.getFullYear(), from.getMonth(), from.getDate()));
+    fromDate.setHours(0, 0, 0, 0);
+    toDate.setUTCHours(toDate.getUTCHours() + 24);
+
+    let entries = await dbHandler.getIndexSorted("diary", "dateTime");
+
+    for (let value of entries) {
+      if (value.dateTime < fromDate || value.dateTime >= toDate) {
+        continue;
       }
 
-      // Make date range inclusive of the whole days at either end
-      let toDate = new Date(Date.UTC(to.getFullYear(), to.getMonth(), to.getDate()));
-      toDate.setHours(0, 0, 0, 0);
-      let fromDate = new Date(Date.UTC(from.getFullYear(), from.getMonth(), from.getDate()));
-      fromDate.setHours(0, 0, 0, 0);
-      toDate.setUTCHours(toDate.getUTCHours() + 24);
+      if (value.items.length > 0 || value.stats.weight != undefined) {
+        result.timestamps.push(value.dateTime);
+        result.items.push(value.items);
+        result.stats.push(value.stats);
+      }
+    }
 
-      dbHandler.getIndex("dateTime", "diary").openCursor(IDBKeyRange.bound(fromDate, toDate, false, true)).onsuccess = function(e) {
-        let cursor = e.target.result;
-
-        if (cursor) {
-
-          let value = cursor.value;
-
-          if (value.items.length > 0 || value.stats.weight != undefined) {
-            result.timestamps.push(value.dateTime);
-            result.items.push(value.items);
-            result.stats.push(value.stats);
-          }
-
-          cursor.continue();
-        } else {
-          resolve(result);
-        }
-      };
-    }).catch(err => {
-      throw (err);
-    });
+    return result;
   },
 
   renderChart: function(data) {
