+++ www/index.html
@@ -115,6 +115,7 @@
   <script src="cordova.js"></script>
   <script src="assets/framework7/framework7-bundle.min.js"></script>
   <script src="assets/js/db-handler.js" defer></script>
+  <script src="assets/js/indexeddb-handler.js" defer></script>
   <script src="assets/js/utils.js" defer></script>
   <script src="assets/js/chartjs/dist/Chart.bundle.min.js"></script>
   <script src="assets/js/chartjs-plugin-annotation/chartjs-plugin-annotation.min.js"></script>
